0.1.1

+++ modules/technical-analysis/src/main/java/com/seomse/trading/Price.java
@@ -16,6 +16,8 @@
 
 package com.seomse.trading;
 
+import java.math.BigDecimal;
+
 /**
  * @author macle
  */
@@ -25,5 +27,5 @@ public interface Price {
      * 종가 얻기
      * @return double 종가
      */
-    double getClose();
+    BigDecimal getClose();
 }
