+++ modules/technical-analysis/src/main/java/com/seomse/trading/PriceChange.java
@@ -17,6 +17,8 @@
 
 package com.seomse.trading;
 
+import java.math.BigDecimal;
+
 /**
  * 가격과 가격 변화
  * @author macle
@@ -30,7 +32,7 @@ public interface PriceChange extends PriceChangeRate, Price{
      * (일별이면 전 거래일, 분봉이면 전봉))
      * @return 변동가격
      */
-    double getChange();
+    BigDecimal getChange();
 
 
 
@@ -38,5 +40,5 @@ public interface PriceChange extends PriceChangeRate, Price{
      * (일별이면 전 거래일, 분봉이면 전봉)
      * @return 전 거래 가격
      */
-    double getPrevious();
+    BigDecimal getPrevious();
 }
