+++ modules/technical-analysis/src/main/java/com/seomse/trading/PriceChangeRate.java
@@ -16,6 +16,8 @@
 
 package com.seomse.trading;
 
+import java.math.BigDecimal;
+
 /**
  * 가격 변화율
  * 가격 변화율만 활용하는 패턴이 있음
@@ -27,5 +29,5 @@ public interface PriceChangeRate {
      * (일별이면 전 거래일, 분봉이면 전봉))
      * @return 변동율
      */
-    double getChangeRate();
+    BigDecimal getChangeRate();
 }
