+++ modules/technical-analysis/src/main/java/com/seomse/trading/technical/analysis/candle/Candle.java
@@ -16,6 +16,8 @@
 
 package com.seomse.trading.technical.analysis.candle;
 
+import java.math.BigDecimal;
+
 /**
  * @author macle
  */
@@ -25,22 +27,22 @@ public interface Candle {
      * 시가 얻기
      * @return 시가 (시작가)
      */
-    double getOpen();
+    BigDecimal getOpen();
 
     /**
      * 종가 얻기
      * @return 종가
      */
-    double getClose();
+    BigDecimal getClose();
 
     /**
      * 고가 얻기
      * @return 고가
      */
-    double getHigh();
+    BigDecimal getHigh();
     /**
      * 저가 얻기
      * @return 저가
      */
-    double getLow();
+    BigDecimal getLow();
 }
